fix: angle detection of custom shapes fluctuates
While aiming at the current direction of the `PlayableCharacter` made
detection consistent, the angle detected could fluctuate based on the
coarseness of the mask found. It was also possible to map to angles that
should be too steep to walk along. Forcing the aim direction to always be
the current mask direction goes some ways towards addressing both
concerns; however, there is a limit to how much more the fluctuations
can be improved due to the inherent fuzziness of detecting surface
angle based only on precise sprite masks.

See also: #14