fix: player does not treat custom masks as walls
When walking or jumping into a custom mask horizontally, wall detection
fails and instead `PlayableCharacter` falls back on floor or ceiling
detection. It appears that the built-in function `collision_line` is not
working properly in this context, so it has been replaced by another
algorithm.

In case the new algorithm results in meaningful performance
issues, the granularity of the detection can be adjusted using the
`portion` variable.

Fixes: #15